+++ ui/temboardui/__main__.py
@@ -31,7 +31,9 @@ def main(*a, **kw):
     # Import main routes
     __import__(__package__ + ".web.routes.core")
     __import__(__package__ + ".web.routes.auth")
+    __import__(__package__ + ".web.routes.inventory")
     __import__(__package__ + ".web.routes.instance")
+    __import__(__package__ + ".web.routes.settings")
 
     # Load sub commands.
     __import__(__package__ + ".cli.apikey")