+++ ui/temboardui/web/routes/core.py
@@ -2,7 +2,7 @@ from flask import current_app as app
 from flask import g, jsonify, redirect
 
 from ...model.orm import Instances
-from ..flask import anonymous_allowed
+from ..flask import admin_required, anonymous_allowed
 
 
 @app.route("/")
@@ -22,3 +22,10 @@ def get_instance_home():
             )
         ]
     )
+
+
+@app.route("/json/plugins")
+@admin_required
+def get_plugins():
+    """List plugins."""
+    return jsonify(sorted(app.temboard.config.temboard.plugins))